requirements.txt
AliyunUtil==0.0.1
backports.ssl-match-hostname==3.4.0.2
Cheetah==2.4.4
cloud-init==0.7.6
configobj==4.7.2
decorator==3.4.0
Django==1.11
django-haystack==2.6.0
django-qiniu-storage==2.3.0
django-redis==4.8.0
iniparse==0.4
jieba==0.38
Jinja2==2.9.5
jsonpatch==1.15
jsonpointer==1.10
Markdown==2.6.8
MarkupSafe==0.23
meld3==1.0.2
MySQL-python==1.2.5
oauth==1.0.1
olefile==0.44
perf==0.1
Pillow==4.1.1
prettytable==0.7.2
pycurl==7.19.0
pygobject==3.14.0
pygpgme==0.3
pyliblzma==0.5.3
pytz==2017.2
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.12
qiniu==7.1.2
redis==2.10.5
requests==2.13.0
six==1.10.0
slip==0.4.0
slip.dbus==0.4.0
supervisor==3.3.1
urlgrabber==3.10
uWSGI==2.0.15
Whoosh==2.7.4
yum-metadata-parser==1.1.4
